Make MiroTransformable output itemsV1 and items in different formats
